Avoid making redundant calls to `ShowResults`
After setting up LSP and Ale, I started seeing a bunch of weird behavior
(flickering cursor, the statusline flipping back and forth between
messages, etc.) that seems to be caused by LSP making multiple calls to
`vim.diagnostic.show` even when the diagnostics aren't actually
changing. I'm not sure if this is a new issue with the Neovim 0.6 API or
already existed with the old one.

This commit updates `set_signs` to only call `ShowResults` once for a
given set of diagnostics, which we achieve by hanging onto a copy in a
buffer-scoped variable and then comparing it against the new data on
each call. This seems to completely resolve the issues I was having.